feat: card creator options & security improvements

+++ src/server/actions/deleteCard.ts
@@ -3,7 +3,25 @@
 import { db } from "~/server/db/";
 import { eq } from "drizzle-orm";
 import { cards as cardSchema } from "~/server/db/schema";
+import { auth } from "@clerk/nextjs/server";
 
-export const deleteCard = async (cardId: string) => {
-  await db.delete(cardSchema).where(eq(cardSchema.id, cardId));
-};
+export async function deleteCard(cardId: string) {
+  const [card] = await db
+    .select()
+    .from(cardSchema)
+    .where(eq(cardSchema.id, cardId));
+
+  if (!card) {
+    return {
+      success: false,
+      message: "Card not found",
+    };
+  }
+
+  const { userId: creatorId } = auth();
+  if (creatorId !== card.creatorId) {
+    return { success: false, message: "Not authorized" };
+  }
+
+  return db.delete(cardSchema).where(eq(cardSchema.id, cardId));
+}
